app/requirements.txt
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
Markdown==3.3.6
pandas==1.4.2
requests==2.27.1
